// src/constants/role-type.ts
export enum RoleTypeWithoutSuperAdmin {
  ADMINISTRATOR = 'ADMINISTRATOR',
  ONCOLOGIST = 'ONCOLOGIST',
  RADIONCOLOGIST = 'RADIONCOLOGIST',
  NURSE = 'NURSE',
  RADIOTHERAPIST = 'RADIOTHERAPIST',
}

export enum RoleType {
  SUPER_ADMINISTRATOR = 'SUPER_ADMINISTRATOR',
  ADMINISTRATOR = RoleTypeWithoutSuperAdmin.ADMINISTRATOR,
  ONCOLOGIST = RoleTypeWithoutSuperAdmin.ONCOLOGIST,
  RADIONCOLOGIST = RoleTypeWithoutSuperAdmin.RADIONCOLOGIST,
  NURSE = RoleTypeWithoutSuperAdmin.NURSE,
  RADIOTHERAPIST = RoleTypeWithoutSuperAdmin.RADIOTHERAPIST,
}